feat: add admin panel and restructure project
