+++ src/app/app.component.html
@@ -1,4 +1,3 @@
 <div class="min-h-screen bg-primary text-bright">
-  <quiz-ui-header />
   <router-outlet />
 </div>
