+++ src/app/pages/admin-panel/admin-panel.component.html
@@ -1,0 +1,2 @@
+<p>admin-panel works!</p>
+<!-- <router-outlet /> -->
