+++ src/app/pages/admin-panel/admin-panel.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'quiz-admin-panel',
+  standalone: true,
+  imports: [],
+  templateUrl: './admin-panel.component.html',
+})
+export class AdminPanelComponent {}
