+++ src/app/pages/quizzes/pages/catalog/catalog.component.ts
@@ -1,16 +1,16 @@
-import { Component, inject, OnInit } from '@angular/core';
-import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
 import { AsyncPipe } from '@angular/common';
-import { Observable, take } from 'rxjs';
-import { UiSpinnerComponent } from '../../shared/ui-kit/ui-spinner/ui-spinner.component';
-import { isLoading$ } from '../../store/categories.store';
-import { UiErrorNotificationComponent } from '../../shared/ui-kit/ui-error-notification/ui-error-notification.component';
-import { QuizCategory } from '../../shared/models/quiz-category.model';
+import { Component, OnInit, inject } from '@angular/core';
 import { Router } from '@angular/router';
-import { CategoriesService } from '../../services/categories-service/categories.service';
-import { CategoriesStoreService } from '../../services/categories-store-service/categories-store.service';
-import { ErrorHandlerService } from '../../services/error-handler-service/error-handler.service';
-import { RandomUtils } from '../../utils/random';
+import { Observable, take } from 'rxjs';
+import { CategoriesService } from '../../../../services/categories-service/categories.service';
+import { CategoriesStoreService } from '../../../../services/categories-store-service/categories-store.service';
+import { ErrorHandlerService } from '../../../../services/error-handler-service/error-handler.service';
+import { QuizCategory } from '../../../../shared/models/quiz-category.model';
+import { UiButtonComponent } from '../../../../shared/ui-kit/ui-button/ui-button.component';
+import { UiErrorNotificationComponent } from '../../../../shared/ui-kit/ui-error-notification/ui-error-notification.component';
+import { UiSpinnerComponent } from '../../../../shared/ui-kit/ui-spinner/ui-spinner.component';
+import { isLoading$ } from '../../../../store/categories.store';
+import { RandomUtils } from '../../../../utils/random';
 import { QuizCardComponent } from './components/quiz-card/quiz-card.component';
 
 @Component({
