+++ src/app/pages/quizzes/pages/catalog/components/quiz-card/quiz-card.constants.ts
