+++ src/app/pages/quizzes/pages/main/main.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
+import { UiButtonComponent } from '../../../../shared/ui-kit/ui-button/ui-button.component';
 
 @Component({
   selector: 'quiz-main',
