+++ src/app/pages/quizzes/pages/quiz/components/question-card/question-card.component.ts
@@ -1,23 +1,23 @@
-import { Component, computed, inject, OnInit, signal } from '@angular/core';
-import { AVATAR_PATHS } from '../../../catalog/components/quiz-card/quiz-card.constants';
-import { UiButtonComponent } from '../../../../shared/ui-kit/ui-button/ui-button.component';
-import { SvgIconComponent } from 'angular-svg-icon';
-import { FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
-import { UiRadioGroupComponent } from '../../../../shared/ui-kit/ui-radio-group/ui-radio-group.component';
-import { ActivatedRoute, Router } from '@angular/router';
-import { Question } from '../../../../shared/models/question.model';
-import { UiSpinnerComponent } from '../../../../shared/ui-kit/ui-spinner/ui-spinner.component';
 import { AsyncPipe } from '@angular/common';
-import { UiErrorNotificationComponent } from '../../../../shared/ui-kit/ui-error-notification/ui-error-notification.component';
+import { Component, OnInit, inject, signal, computed } from '@angular/core';
+import { ReactiveFormsModule, FormControl, Validators } from '@angular/forms';
+import { Router, ActivatedRoute } from '@angular/router';
+import { SvgIconComponent } from 'angular-svg-icon';
 import { Subscription } from 'rxjs';
-import { QuestionsService } from '../../../../services/questions-service/questions.service';
-import { ErrorHandlerService } from '../../../../services/error-handler-service/error-handler.service';
-import { DialogService } from '../../../../services/dialog-service/dialog.service';
-import { StatisticsService } from '../../../../services/statistics-service/statistics.service';
+import { DialogService } from '../../../../../../services/dialog-service/dialog.service';
+import { ErrorHandlerService } from '../../../../../../services/error-handler-service/error-handler.service';
+import { QuestionsService } from '../../../../../../services/questions-service/questions.service';
+import { StatisticsService } from '../../../../../../services/statistics-service/statistics.service';
+import { Question } from '../../../../../../shared/models/question.model';
+import { UiButtonComponent } from '../../../../../../shared/ui-kit/ui-button/ui-button.component';
+import { UiErrorNotificationComponent } from '../../../../../../shared/ui-kit/ui-error-notification/ui-error-notification.component';
 import {
   defaultDialog,
   finishQuizDialog,
-} from '../../../../shared/ui-kit/ui-navigation-confirm-dialog/dialog-states';
+} from '../../../../../../shared/ui-kit/ui-navigation-confirm-dialog/dialog-states';
+import { UiRadioGroupComponent } from '../../../../../../shared/ui-kit/ui-radio-group/ui-radio-group.component';
+import { UiSpinnerComponent } from '../../../../../../shared/ui-kit/ui-spinner/ui-spinner.component';
+import { AVATAR_PATHS } from '../../../catalog/components/quiz-card/quiz-card.constants';
 
 @Component({
   selector: 'quiz-question-card',
