+++ src/app/pages/quizzes/pages/quiz/quiz.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { QuestionCardComponent } from './components/question-card/question-card.component';
-import { UiButtonComponent } from '../../shared/ui-kit/ui-button/ui-button.component';
-import { NavigationConfirmDialogComponent } from '../../shared/ui-kit/ui-navigation-confirm-dialog/ui-navigation-confirm-dialog.component';
+import { UiButtonComponent } from '../../../../shared/ui-kit/ui-button/ui-button.component';
+import { NavigationConfirmDialogComponent } from '../../../../shared/ui-kit/ui-navigation-confirm-dialog/ui-navigation-confirm-dialog.component';
 
 @Component({
   selector: 'quiz-quiz',
