+++ src/app/pages/quizzes/quizzes.component.html
@@ -1,0 +1,2 @@
+<quiz-ui-header />
+<router-outlet />
