+++ src/app/shared/ui-kit/ui-header/ui-header.component.html
@@ -24,7 +24,9 @@
   </div>
 
   <div class="hidden items-center space-x-8 md:flex">
-    <quiz-ui-button variant="accent"> Login </quiz-ui-button>
+    <quiz-ui-button variant="accent" route="admin">
+      Admin panel
+    </quiz-ui-button>
   </div>
 
   <!-- Mobile Menu Button -->
@@ -65,7 +67,9 @@
         }
       </ul>
       <div class="mt-10 flex items-center">
-        <quiz-ui-button variant="accent"> Login </quiz-ui-button>
+        <quiz-ui-button variant="accent" route="admin">
+          Admin panel</quiz-ui-button
+        >
       </div>
     </div>
   </div>
